Student emotion monitoring/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.1.0
PyYAML>=6.0
jupyter>=1.0.0
notebook>=6.4.0
chardet>=4.0.0